Extracted the do_set from the act.wizard.cpp and the DB.
