+++ src/act_wizard.cpp
@@ -113,12 +113,10 @@ void log_zone_count_reset();
 void appear(CharData *ch);
 void ResetZone(ZoneRnum zone);
 //extern CharData *find_char(long n);
-void RenamePlayer(CharData *ch, char *oname);
 int _parse_name(char *arg, char *name);
 int Valid_Name(char *name);
 int reserved_word(const char *name);
 // local functions
-int perform_set(CharData *ch, CharData *vict, int mode, char *val_arg);
 void perform_immort_invis(CharData *ch, int level);
 void do_echo(CharData *ch, char *argument, int cmd, int subcmd);
 void do_send(CharData *ch, char *argument, int cmd, int subcmd);
@@ -150,7 +148,6 @@ void do_wiznet(CharData *ch, char *argument, int cmd, int subcmd);
 void do_zreset(CharData *ch, char *argument, int cmd, int subcmd);
 void do_wizutil(CharData *ch, char *argument, int cmd, int subcmd);
 void do_show(CharData *ch, char *argument, int cmd, int subcmd);
-void do_set(CharData *ch, char *argument, int cmd, int subcmd);
 void do_liblist(CharData *ch, char *argument, int cmd, int subcmd);
 //
 void do_sdemigod(CharData *ch, char *argument, int cmd, int subcmd);
@@ -231,8 +228,6 @@ void send_to_gods(char *text, bool demigod) {
 		}
 	}
 }
-
-#define MAX_TIME 0x7fffffff
 
 extern const char *deaf_social;
 
@@ -452,7 +447,7 @@ int set_punish(CharData *ch, CharData *vict, int punish, char *reason, long time
 	}
 	assert(pundata);
 	if (GetRealLevel(ch) < pundata->level) {
-		SendMsgToChar("Да кто ты такой!!? Чтобы оспаривать волю СТАРШИХ БОГОВ !!!\r\n", ch);
+		SendMsgToChar("Да кто ты такой?!! Чтобы оспаривать волю СТАРШИХ БОГОВ!!!\r\n", ch);
 		return 0;
 	}
 	// Проверяем наказание или освобождение.
@@ -2813,952 +2808,6 @@ void show_apply(CharData *ch, CharData *vict) {
 			}
 		}
 	}
-}
-
-// **************** The do_set function
-
-#define PC   1
-#define NPC  2
-#define BOTH 3
-
-#define MISC    0
-#define BINARY    1
-#define NUMBER    2
-
-inline void SET_OR_REMOVE(const bool on, const bool off, FlagData &flagset, const Bitvector packed_flag) {
-	if (on) {
-		flagset.set(packed_flag);
-	} else if (off) {
-		flagset.unset(packed_flag);
-	}
-}
-
-#define RANGE(low, high) (value = MAX((low), MIN((high), (value))))
-
-// The set options available
-struct set_struct        /*
-				   { const char *cmd;
-				   const char level;
-				   const char pcnpc;
-				   const char type;
-				   } */ set_fields[] =
-	{
-		{"brief", kLvlGod, PC, BINARY},    // 0
-		{"invstart", kLvlGod, PC, BINARY},    // 1
-		{"nosummon", kLvlGreatGod, PC, BINARY},
-		{"maxhit", kLvlImplementator, BOTH, NUMBER},
-		{"maxmana", kLvlGreatGod, BOTH, NUMBER},
-		{"maxmove", kLvlImplementator, BOTH, NUMBER},    // 5
-		{"hit", kLvlGreatGod, BOTH, NUMBER},
-		{"mana", kLvlGreatGod, BOTH, NUMBER},
-		{"move", kLvlGreatGod, BOTH, NUMBER},
-		{"race", kLvlGreatGod, BOTH, NUMBER},
-		{"size", kLvlImplementator, BOTH, NUMBER},    // 10
-		{"ac", kLvlGreatGod, BOTH, NUMBER},
-		{"gold", kLvlImplementator, BOTH, NUMBER},
-		{"bank", kLvlImplementator, PC, NUMBER},
-		{"exp", kLvlImplementator, BOTH, NUMBER},
-		{"hitroll", kLvlImplementator, BOTH, NUMBER}, // 15
-		{"damroll", kLvlImplementator, BOTH, NUMBER},
-		{"invis", kLvlImplementator, PC, NUMBER},
-		{"nohassle", kLvlImplementator, PC, BINARY},
-		{"frozen", kLvlGreatGod, PC, MISC},
-		{"practices", kLvlGreatGod, PC, NUMBER}, // 20
-		{"lessons", kLvlGreatGod, PC, NUMBER},
-		{"drunk", kLvlGreatGod, BOTH, MISC},
-		{"hunger", kLvlGreatGod, BOTH, MISC},
-		{"thirst", kLvlGreatGod, BOTH, MISC},
-		{"thief", kLvlGod, PC, BINARY}, // 25
-		{"level", kLvlImplementator, BOTH, NUMBER},
-		{"room", kLvlImplementator, BOTH, NUMBER},
-		{"roomflag", kLvlGreatGod, PC, BINARY},
-		{"siteok", kLvlGreatGod, PC, BINARY},
-		{"deleted", kLvlImplementator, PC, BINARY}, // 30
-		{"class", kLvlImplementator, BOTH, MISC},
-		{"demigod", kLvlImplementator, PC, BINARY},
-		{"loadroom", kLvlGreatGod, PC, MISC},
-		{"color", kLvlGod, PC, BINARY},
-		{"idnum", kLvlImplementator, PC, NUMBER}, // 35
-		{"passwd", kLvlImplementator, PC, MISC},
-		{"nodelete", kLvlGod, PC, BINARY},
-		{"sex", kLvlGreatGod, BOTH, MISC},
-		{"age", kLvlGreatGod, BOTH, NUMBER},
-		{"height", kLvlGod, BOTH, NUMBER}, // 40
-		{"weight", kLvlGod, BOTH, NUMBER},
-		{"godslike", kLvlImplementator, BOTH, BINARY},
-		{"godscurse", kLvlImplementator, BOTH, BINARY},
-		{"olc", kLvlImplementator, PC, NUMBER},
-		{"name", kLvlGreatGod, PC, MISC}, // 45
-		{"trgquest", kLvlImplementator, PC, MISC},
-		{"mkill", kLvlImplementator, PC, MISC},
-		{"highgod", kLvlImplementator, PC, MISC},
-		{"hell", kLvlGod, PC, MISC},
-		{"email", kLvlGod, PC, MISC}, //50
-		{"religion", kLvlGod, PC, MISC},
-		{"perslog", kLvlImplementator, PC, BINARY},
-		{"mute", kLvlGod, PC, MISC},
-		{"dumb", kLvlGod, PC, MISC},
-		{"karma", kLvlImplementator, PC, MISC},
-		{"unreg", kLvlGod, PC, MISC}, // 56
-		{"палач", kLvlImplementator, PC, BINARY}, // 57
-		{"killer", kLvlImplementator, PC, BINARY}, // 58
-		{"remort", kLvlImplementator, PC, NUMBER}, // 59
-		{"tester", kLvlImplementator, PC, BINARY}, // 60
-		{"autobot", kLvlImplementator, PC, BINARY}, // 61
-		{"hryvn", kLvlImplementator, PC, NUMBER}, // 62
-		{"scriptwriter", kLvlImplementator, PC, BINARY}, // 63
-		{"spammer", kLvlGod, PC, BINARY}, // 64
-		{"gloryhide", kLvlImplementator, PC, BINARY}, // 65
-		{"telegram", kLvlImplementator, PC, MISC}, // 66
-		{"nogata", kLvlImplementator, PC, NUMBER}, // 67
-		{"position", kLvlImplementator, PC, NUMBER},
-		{"skilltester", kLvlImplementator, PC, BINARY}, //69
-		{"quest", kLvlImplementator, PC, NUMBER}, //70
-		{"\n", 0, BOTH, MISC}
-	};
-
-int perform_set(CharData *ch, CharData *vict, int mode, char *val_arg) {
-	int i, j, c, value = 0, return_code = 1, ptnum, times = 0;
-	bool on = false;
-	bool off = false;
-	char npad[ECase::kLastCase + 1][256];
-	char *reason;
-	RoomRnum rnum;
-	RoomVnum rvnum;
-	char output[kMaxStringLength], num[kMaxInputLength];
-	int rod;
-
-	// Check to make sure all the levels are correct
-	if (!IS_IMPL(ch)) {
-		if (!vict->IsNpc() && vict != ch) {
-			if (!GET_GOD_FLAG(ch, EGf::kDemigod)) {
-				if (GetRealLevel(ch) <= GetRealLevel(vict) && !PRF_FLAGGED(ch, EPrf::kCoderinfo)) {
-					SendMsgToChar("Это не так просто, как вам кажется...\r\n", ch);
-					return (0);
-				}
-			} else {
-				if (GetRealLevel(vict) >= kLvlImmortal || PRF_FLAGGED(vict, EPrf::kCoderinfo)) {
-					SendMsgToChar("Это не так просто, как вам кажется...\r\n", ch);
-					return (0);
-				}
-			}
-		}
-	}
-	if (!privilege::HasPrivilege(ch, std::string(set_fields[mode].cmd), 0, 1)) {
-		SendMsgToChar("Кем вы себя возомнили?\r\n", ch);
-		return (0);
-	}
-
-	// Make sure the PC/NPC is correct
-	if (vict->IsNpc() && !(set_fields[mode].pcnpc & NPC)) {
-		SendMsgToChar("Эта тварь недостойна такой чести!\r\n", ch);
-		return (0);
-	} else if (!vict->IsNpc() && !(set_fields[mode].pcnpc & PC)) {
-		act("Вы оскорбляете $S - $E ведь не моб!", false, ch, nullptr, vict, kToChar);
-		return (0);
-	}
-
-	// Find the value of the argument
-	if (set_fields[mode].type == BINARY) {
-		if (!strn_cmp(val_arg, "on", 2) || !strn_cmp(val_arg, "yes", 3) || !strn_cmp(val_arg, "вкл", 3))
-			on = true;
-		else if (!strn_cmp(val_arg, "off", 3) || !strn_cmp(val_arg, "no", 2) || !strn_cmp(val_arg, "выкл", 4))
-			off = true;
-		if (!(on || off)) {
-			SendMsgToChar("Значение может быть 'on' или 'off'.\r\n", ch);
-			return (0);
-		}
-		sprintf(output, "%s %s для %s.", set_fields[mode].cmd, ONOFF(on), GET_PAD(vict, 1));
-	} else if (set_fields[mode].type == NUMBER) {
-		value = atoi(val_arg);
-		sprintf(output, "У %s %s установлено в %d.", GET_PAD(vict, 1), set_fields[mode].cmd, value);
-	} else {
-		strcpy(output, "Хорошо.");
-	}
-	switch (mode) {
-		case 0: SET_OR_REMOVE(on, off, PRF_FLAGS(vict), EPrf::kBrief);
-			break;
-		case 1: SET_OR_REMOVE(on, off, PLR_FLAGS(vict), EPlrFlag::kInvStart);
-			break;
-		case 2: SET_OR_REMOVE(on, off, PRF_FLAGS(vict), EPrf::KSummonable);
-			sprintf(output, "Возможность призыва %s для %s.\r\n", ONOFF(!on), GET_PAD(vict, 1));
-			break;
-		case 3: vict->points.max_hit = RANGE(1, 5000);
-			affect_total(vict);
-			break;
-		case 4: break;
-		case 5: vict->points.max_move = RANGE(1, 5000);
-			affect_total(vict);
-			break;
-		case 6: vict->points.hit = RANGE(-9, vict->points.max_hit);
-			affect_total(vict);
-			update_pos(vict);
-			break;
-		case 7: break;
-		case 8: break;
-		case 9:
-			// Выставляется род для РС
-			rod = PlayerRace::CheckRace(GET_KIN(ch), val_arg);
-			if (rod == RACE_UNDEFINED) {
-				SendMsgToChar("Не было таких на земле русской!\r\n", ch);
-				SendMsgToChar(PlayerRace::ShowRacesMenu(GET_KIN(ch)), ch);
-				return (0);
-			} else {
-				GET_RACE(vict) = rod;
-				affect_total(vict);
-
-			}
-			break;
-		case 10: vict->real_abils.size = RANGE(1, 100);
-			affect_total(vict);
-			break;
-		case 11: vict->real_abils.armor = RANGE(-100, 100);
-			affect_total(vict);
-			break;
-		case 12: vict->set_gold(value);
-			break;
-		case 13: vict->set_bank(value);
-			break;
-		case 14:
-			//vict->points.exp = RANGE(0, 7000000);
-			RANGE(0, GetExpUntilNextLvl(vict, kLvlImmortal) - 1);
-			gain_exp_regardless(vict, value - GET_EXP(vict));
-			break;
-		case 15: vict->real_abils.hitroll = RANGE(-20, 20);
-			affect_total(vict);
-			break;
-		case 16: vict->real_abils.damroll = RANGE(-20, 20);
-			affect_total(vict);
-			break;
-		case 17:
-			if (!IS_IMPL(ch) && ch != vict && !PRF_FLAGGED(ch, EPrf::kCoderinfo)) {
-				SendMsgToChar("Вы не столь Божественны, как вам кажется!\r\n", ch);
-				return (0);
-			}
-			SET_INVIS_LEV(vict, RANGE(0, GetRealLevel(vict)));
-			break;
-		case 18:
-			if (!IS_IMPL(ch) && ch != vict && !PRF_FLAGGED(ch, EPrf::kCoderinfo)) {
-				SendMsgToChar("Вы не столь Божественны, как вам кажется!\r\n", ch);
-				return (0);
-			}
-			SET_OR_REMOVE(on, off, PRF_FLAGS(vict), EPrf::kNohassle);
-			break;
-		case 19: reason = one_argument(val_arg, num);
-			if (!*num) {
-				SendMsgToChar(ch, "Укажите срок наказания.\r\n");
-				return (0); // return(0) обходит запись в файл в do_set
-			}
-			if (!strcmp(num, "0")) {
-				if (!set_punish(ch, vict, SCMD_FREEZE, reason, times)) {
-					return (0);
-				}
-				return (1);
-			}
-			times = atol(num);
-			if (times == 0) {
-				SendMsgToChar(ch, "Срок указан не верно.\r\n");
-				return (0);
-			} else {
-				if (!set_punish(ch, vict, SCMD_FREEZE, reason, times))
-					return (0);
-			}
-			break;
-		case 20:
-		case 21: return_code = 0;
-			break;
-		case 22:
-		case 23:
-		case 24: {
-			const unsigned num = mode - 22; // magic number циркулевских времен
-			if (num >= (ch)->player_specials->saved.conditions.size()) {
-				SendMsgToChar("Ошибка: num >= saved.conditions.size(), сообщите кодерам.\r\n", ch);
-				return 0;
-			}
-			if (!str_cmp(val_arg, "off") || !str_cmp(val_arg, "выкл")) {
-				GET_COND(vict, num) = -1;
-				sprintf(output, "Для %s %s сейчас отключен.", GET_PAD(vict, 1), set_fields[mode].cmd);
-			} else if (is_number(val_arg)) {
-				value = atoi(val_arg);
-				RANGE(0, kMaxCondition);
-				GET_COND(vict, num) = value;
-				sprintf(output, "Для %s %s установлен в %d.", GET_PAD(vict, 1), set_fields[mode].cmd, value);
-			} else {
-				SendMsgToChar("Должно быть 'off' или значение от 0 до 24.\r\n", ch);
-				return 0;
-			}
-			break;
-		}
-		case 25: SET_OR_REMOVE(on, off, PLR_FLAGS(vict), EPlrFlag::kBurglar);
-			break;
-		case 26:
-			if (!PRF_FLAGGED(ch, EPrf::kCoderinfo)
-				&& (value > GetRealLevel(ch) || value > kLvlImplementator || GetRealLevel(vict) > GetRealLevel(ch))) {
-				SendMsgToChar("Вы не можете установить уровень игрока выше собственного.\r\n", ch);
-				return (0);
-			}
-			RANGE(0, kLvlImplementator);
-			vict->set_level(value);
-			break;
-		case 27:
-			if ((rnum = GetRoomRnum(value)) == kNowhere) {
-				SendMsgToChar("Поищите другой МУД. В этом МУДе нет такой комнаты.\r\n", ch);
-				return (0);
-			}
-			if (IN_ROOM(vict) != kNowhere)    // Another Eric Green special.
-				RemoveCharFromRoom(vict);
-			PlaceCharToRoom(vict, rnum);
-			vict->dismount();
-			break;
-		case 28: SET_OR_REMOVE(on, off, PRF_FLAGS(vict), EPrf::kRoomFlags);
-			break;
-		case 29: SET_OR_REMOVE(on, off, PLR_FLAGS(vict), EPlrFlag::kSiteOk);
-			break;
-		case 30:
-			if (IS_IMPL(vict) || PRF_FLAGGED(vict, EPrf::kCoderinfo)) {
-				SendMsgToChar("Истинные боги вечны!\r\n", ch);
-				return 0;
-			}
-			SET_OR_REMOVE(on, off, PLR_FLAGS(vict), EPlrFlag::kDeleted);
-			if (PLR_FLAGS(vict).get(EPlrFlag::kDeleted)) {
-				if (PLR_FLAGS(vict).get(EPlrFlag::kNoDelete)) {
-					PLR_FLAGS(vict).unset(EPlrFlag::kNoDelete);
-					SendMsgToChar("NODELETE flag also removed.\r\n", ch);
-				}
-			}
-			break;
-		case 31: {
-			auto class_id = FindAvailableCharClassId(val_arg);
-			if (class_id == ECharClass::kUndefined) {
-				SendMsgToChar("Нет такого класса в этой игре. Найдите себе другую.\r\n", ch);
-				return (0);
-			}
-			vict->set_class(class_id);
-			break;
-		}
-		case 32:
-			// Флаг для морталов с привилегиями
-			if (!IS_IMPL(ch) && !PRF_FLAGGED(ch, EPrf::kCoderinfo)) {
-				SendMsgToChar("Вы не столь Божественны, как вам кажется!\r\n", ch);
-				return 0;
-			}
-			if (on) {
-				SET_GOD_FLAG(vict, EGf::kDemigod);
-			} else if (off) {
-				CLR_GOD_FLAG(vict, EGf::kDemigod);
-			}
-			break;
-		case 33:
-			if (is_number(val_arg)) {
-				rvnum = atoi(val_arg);
-				if (GetRoomRnum(rvnum) != kNowhere) {
-					GET_LOADROOM(vict) = rvnum;
-					sprintf(output, "%s будет входить в игру из комнаты #%d.",
-							GET_NAME(vict), GET_LOADROOM(vict));
-				} else {
-					SendMsgToChar
-						("Прежде чем кого-то куда-то поместить, надо это КУДА-ТО создать.\r\n"
-						 "Скажите Стрибогу - пусть зоны рисует, а не пьянствует.\r\n", ch);
-					return (0);
-				}
-			} else {
-				SendMsgToChar("Должен быть виртуальный номер комнаты.\r\n", ch);
-				return (0);
-			}
-			break;
-		case 34: SET_OR_REMOVE(on, off, PRF_FLAGS(vict), EPrf::kColor1);
-			SET_OR_REMOVE(on, off, PRF_FLAGS(vict), EPrf::kColor2);
-			break;
-		case 35:
-			if (!IS_IMPL(ch) || !vict->IsNpc()) {
-				return (0);
-			}
-			vict->set_idnum(value);
-			break;
-		case 36:
-			if (!IS_IMPL(ch) && !PRF_FLAGGED(ch, EPrf::kCoderinfo) && ch != vict) {
-				SendMsgToChar("Давайте не будем экспериментировать.\r\n", ch);
-				return (0);
-			}
-			if (IS_IMPL(vict) && ch != vict) {
-				SendMsgToChar("Вы не можете ЭТО изменить.\r\n", ch);
-				return (0);
-			}
-			if (!Password::check_password(vict, val_arg)) {
-				SendMsgToChar(ch, "%s\r\n", Password::BAD_PASSWORD);
-				return 0;
-			}
-			Password::set_password(vict, val_arg);
-			Password::send_password(GET_EMAIL(vict), val_arg, std::string(GET_NAME(vict)));
-			sprintf(buf, "%s заменен пароль богом.", GET_PAD(vict, 2));
-			AddKarma(vict, buf, GET_NAME(ch));
-			sprintf(output, "Пароль изменен на '%s'.", val_arg);
-			break;
-		case 37: SET_OR_REMOVE(on, off, PLR_FLAGS(vict), EPlrFlag::kNoDelete);
-			break;
-		case 38:
-			if ((i = search_block(val_arg, genders, false)) < 0) {
-				SendMsgToChar
-					("Может быть 'мужчина', 'женщина', или 'бесполое'(а вот это я еще не оценил :).\r\n", ch);
-				return (0);
-			}
-			vict->set_sex(static_cast<EGender>(i));
-			break;
-
-		case 39:        // set age
-			if (value < 2 || value > 200)    // Arbitrary limits.
-			{
-				SendMsgToChar("Поддерживаются возрасты от 2 до 200.\r\n", ch);
-				return (0);
-			}
-			/*
-		 * NOTE: May not display the exact age specified due to the integer
-		 * division used elsewhere in the code.  Seems to only happen for
-		 * some values below the starting age (17) anyway. -gg 5/27/98
-		 */
-			vict->player_data.time.birth = time(nullptr) - ((value - 17) * kSecsPerMudYear);
-			break;
-
-		case 40:        // Blame/Thank Rick Glover. :)
-			GET_HEIGHT(vict) = value;
-			affect_total(vict);
-			break;
-
-		case 41: GET_WEIGHT(vict) = value;
-			affect_total(vict);
-			break;
-
-		case 42:
-			if (on) {
-				SET_GOD_FLAG(vict, EGf::kGodsLike);
-				if (sscanf(val_arg, "%s %d", npad[0], &i) != 0)
-					GCURSE_DURATION(vict) = (i > 0) ? time(nullptr) + i * 60 * 60 : MAX_TIME;
-				else
-					GCURSE_DURATION(vict) = 0;
-				sprintf(buf, "%s установил GUDSLIKE персонажу %s.", GET_NAME(ch), GET_NAME(vict));
-				mudlog(buf, BRF, kLvlImplementator, SYSLOG, 0);
-
-			} else if (off)
-				CLR_GOD_FLAG(vict, EGf::kGodsLike);
-			break;
-		case 43:
-			if (on) {
-				SET_GOD_FLAG(vict, EGf::kGodscurse);
-				if (sscanf(val_arg, "%s %d", npad[0], &i) != 0)
-					GCURSE_DURATION(vict) = (i > 0) ? time(nullptr) + i * 60 * 60 : MAX_TIME;
-				else
-					GCURSE_DURATION(vict) = 0;
-			} else if (off)
-				CLR_GOD_FLAG(vict, EGf::kGodscurse);
-			break;
-		case 44:
-			if (PRF_FLAGGED(ch, EPrf::kCoderinfo) || IS_IMPL(ch))
-				GET_OLC_ZONE(vict) = value;
-			else {
-				sprintf(buf, "Слишком низкий уровень чтоб раздавать права OLC.\r\n");
-				SendMsgToChar(buf, ch);
-			}
-			break;
-
-		case 45:
-			// изменение имени !!!
-
-			if ((i = sscanf(val_arg, "%s %s %s %s %s %s", npad[0], npad[1], npad[2], npad[3], npad[4], npad[5])) != 6) {
-				sprintf(buf, "Требуется указать 6 падежей, найдено %d\r\n", i);
-				SendMsgToChar(buf, ch);
-				return (0);
-			}
-
-			if (*npad[0] == '*')    // Only change pads
-			{
-				for (i = ECase::kGen; i <= ECase::kLastCase; i++)
-					if (!_parse_name(npad[i], npad[i])) {
-						vict->player_data.PNames[i] = std::string(npad[i]);
-					}
-				sprintf(buf, "Произведена замена падежей.\r\n");
-				SendMsgToChar(buf, ch);
-			} else {
-				for (i = ECase::kFirstCase; i <= ECase::kLastCase; i++) {
-					if (strlen(npad[i]) < kMinNameLength || strlen(npad[i]) > kMaxNameLength) {
-						sprintf(buf, "Падеж номер %d некорректен.\r\n", ++i);
-						SendMsgToChar(buf, ch);
-						return (0);
-					}
-				}
-
-				if (_parse_name(npad[0], npad[0]) ||
-					strlen(npad[0]) < kMinNameLength ||
-					strlen(npad[0]) > kMaxNameLength ||
-					!Valid_Name(npad[0]) || reserved_word(npad[0]) || fill_word(npad[0])) {
-					SendMsgToChar("Некорректное имя.\r\n", ch);
-					return (0);
-				}
-
-				if ((GetPlayerIdByName(npad[0]) >= 0) && !PLR_FLAGS(vict).get(EPlrFlag::kDeleted)) {
-					SendMsgToChar("Это имя совпадает с именем другого персонажа.\r\n"
-								  "Для исключения различного рода недоразумений имя отклонено.\r\n", ch);
-					return (0);
-				}
-				// выносим из листа неодобренных имен, если есть
-				NewNames::remove(vict);
-
-				ptnum = GetPlayerTablePosByName(GET_NAME(vict));
-				if (ptnum < 0)
-					return (0);
-
-				if (!PLR_FLAGS(vict).get(EPlrFlag::kFrozen)
-					&& !PLR_FLAGS(vict).get(EPlrFlag::kDeleted)
-					&& !IS_IMMORTAL(vict)) {
-					TopPlayer::Remove(vict);
-				}
-
-				for (i = ECase::kFirstCase; i <= ECase::kLastCase; i++) {
-					if (!_parse_name(npad[i], npad[i])) {
-						vict->player_data.PNames[i] = std::string(npad[i]);
-					}
-				}
-				sprintf(buf, "Name changed from %s to %s", GET_NAME(vict), npad[0]);
-				vict->set_name(npad[0]);
-				AddKarma(vict, buf, GET_NAME(ch));
-
-				if (!PLR_FLAGS(vict).get(EPlrFlag::kFrozen)
-					&& !PLR_FLAGS(vict).get(EPlrFlag::kDeleted)
-					&& !IS_IMMORTAL(vict)) {
-					TopPlayer::Refresh(vict);
-				}
-
-				player_table.SetName(ptnum, npad[0]);
-
-				return_code = 2;
-				PLR_FLAGS(vict).set(EPlrFlag::kCrashSave);
-			}
-			break;
-
-		case 46:
-
-			npad[1][0] = '\0';
-			if (sscanf(val_arg, "%d %s %[^\n]", &ptnum, npad[0], npad[1]) != 3) {
-				if (sscanf(val_arg, "%d %s", &ptnum, npad[0]) != 2) {
-					SendMsgToChar("Формат : set <имя> trgquest <quest_num> <on|off> <строка данных>\r\n", ch);
-					return 0;
-				}
-			}
-
-			if (!str_cmp(npad[0], "off") || !str_cmp(npad[0], "выкл")) {
-				if (!vict->quested_remove(ptnum)) {
-					act("$N не выполнял$G этого квеста.", false, ch, nullptr, vict, kToChar);
-					return 0;
-				}
-			} else if (!str_cmp(npad[0], "on") || !str_cmp(npad[0], "вкл")) {
-				vict->quested_add(vict, ptnum, npad[1]);
-			} else {
-				SendMsgToChar("Требуется on или off.\r\n", ch);
-				return 0;
-			}
-			break;
-
-		case 47:
-
-			if (sscanf(val_arg, "%d %s", &ptnum, npad[0]) != 2) {
-				SendMsgToChar("Формат : set <имя> mkill <MobVnum> <off|num>\r\n", ch);
-				return (0);
-			}
-			if (!str_cmp(npad[0], "off") || !str_cmp(npad[0], "выкл"))
-				vict->mobmax_remove(ptnum);
-			else if ((j = atoi(npad[0])) > 0) {
-				if ((c = vict->mobmax_get(ptnum)) != j)
-					vict->mobmax_add(vict, ptnum, j - c, MobMax::get_level_by_vnum(ptnum));
-				else {
-					act("$N убил$G именно столько этих мобов.", false, ch, nullptr, vict, kToChar);
-					return (0);
-				}
-			} else {
-				SendMsgToChar("Требуется off или значение больше 0.\r\n", ch);
-				return (0);
-			}
-			break;
-
-		case 48: return (0);
-			break;
-
-		case 49: reason = one_argument(val_arg, num);
-			if (!*num) {
-				SendMsgToChar(ch, "Укажите срок наказания.\r\n");
-				return (0); // return(0) обходит запись в файл в do_set
-			}
-			if (!str_cmp(num, "0")) {
-				if (!set_punish(ch, vict, SCMD_HELL, reason, times)) {
-					return (0);
-				}
-				return (1);
-			}
-			times = atol(num);
-			if (times == 0) {
-				SendMsgToChar(ch, "Срок указан не верно.\r\n");
-				return (0);
-			} else {
-				if (!set_punish(ch, vict, SCMD_HELL, reason, times))
-					return (0);
-			}
-			break;
-		case 50:
-			if (IsValidEmail(val_arg)) {
-				utils::ConvertToLow(val_arg);
-				sprintf(buf, "Email changed from %s to %s", GET_EMAIL(vict), val_arg);
-				AddKarma(vict, buf, GET_NAME(ch));
-				strncpy(GET_EMAIL(vict), val_arg, 127);
-				*(GET_EMAIL(vict) + 127) = '\0';
-			} else {
-				SendMsgToChar("Wrong E-Mail.\r\n", ch);
-				return (0);
-			}
-			break;
-
-		case 51:
-			// Выставляется род для РС
-			rod = (*val_arg);
-			if (rod != '0' && rod != '1') {
-				SendMsgToChar("Не было таких на земле русской!\r\n", ch);
-				SendMsgToChar("0 - Язычество, 1 - Христианство\r\n", ch);
-				return (0);
-			} else {
-				GET_RELIGION(vict) = rod - '0';
-			}
-			break;
-
-		case 52:
-			// Отдельный лог команд персонажа
-			if (on) {
-				SET_GOD_FLAG(vict, EGf::kPerslog);
-			} else if (off) {
-				CLR_GOD_FLAG(vict, EGf::kPerslog);
-			}
-			break;
-
-		case 53: reason = one_argument(val_arg, num);
-			if (!*num) {
-				SendMsgToChar(ch, "Укажите срок наказания.\r\n");
-				return (0); // return(0) обходит запись в файл в do_set
-			}
-			if (!str_cmp(num, "0")) {
-				if (!set_punish(ch, vict, SCMD_MUTE, reason, times)) {
-					return (0);
-				}
-				return (1);
-			}
-			times = atol(num);
-			if (times == 0) {
-				SendMsgToChar(ch, "Срок указан не верно.\r\n");
-				return (0);
-			} else {
-				if (!set_punish(ch, vict, SCMD_MUTE, reason, times))
-					return (0);
-			}
-			break;
-
-		case 54: reason = one_argument(val_arg, num);
-			if (!*num) {
-				SendMsgToChar(ch, "Укажите срок наказания.\r\n");
-				return (0); // return(0) обходит запись в файл в do_set
-			}
-			if (!str_cmp(num, "0")) {
-				if (!set_punish(ch, vict, SCMD_DUMB, reason, times)) {
-					return (0);
-				}
-				return (1);
-			}
-			times = atol(num);
-			if (times == 0) {
-				SendMsgToChar(ch, "Срок указан не верно.\r\n");
-				return (0);
-			} else {
-				if (!set_punish(ch, vict, SCMD_DUMB, reason, times))
-					return (0);
-			}
-			break;
-		case 55:
-			if (GetRealLevel(vict) >= kLvlImmortal && !IS_IMPL(ch) && !PRF_FLAGGED(ch, EPrf::kCoderinfo)) {
-				SendMsgToChar("Кем вы себя возомнили?\r\n", ch);
-				return 0;
-			}
-			reason = strdup(val_arg);
-			if (reason && *reason) {
-				skip_spaces(&reason);
-				sprintf(buf, "add by %s", GET_NAME(ch));
-				if (!strcmp(reason, "clear")) {
-					if KARMA(vict)
-						free(KARMA(vict));
-
-					KARMA(vict) = nullptr;
-					act("Вы отпустили $N2 все грехи.", false, ch, nullptr, vict, kToChar);
-					sprintf(buf, "%s", GET_NAME(ch));
-					AddKarma(vict, "Очистка грехов", buf);
-
-				} else AddKarma(vict, buf, reason);
-			} else {
-				SendMsgToChar("Формат команды: set [ file | player ] <character> karma <reason>\r\n", ch);
-				return (0);
-			}
-			break;
-
-		case 56:      // Разрегистрация персонажа
-			reason = one_argument(val_arg, num);
-			if (!set_punish(ch, vict, SCMD_UNREGISTER, reason, 0)) return (0);
-			break;
-
-		case 57:      // Установка флага палач
-			reason = one_argument(val_arg, num);
-			skip_spaces(&reason);
-			sprintf(buf, "executor %s by %s", (on ? "on" : "off"), GET_NAME(ch));
-//			AddKarma(vict, buf, reason);
-			if (on) {
-				PRF_FLAGS(vict).set(EPrf::kExecutor);
-			} else if (off) {
-				PRF_FLAGS(vict).unset(EPrf::kExecutor);
-			}
-			break;
-
-		case 58: // Снятие или постановка флага !ДУШЕГУБ! только для имплементоров
-			SET_OR_REMOVE(on, off, PLR_FLAGS(vict), EPlrFlag::kKiller);
-			break;
-		case 59: // флаг реморта
-			if (value > 1 && value < 75) {
-				sprintf(buf, "Иммортал %s установил реморт %d  для игрока %s ", GET_NAME(ch), value, GET_NAME(vict));
-				AddKarma(vict, buf, GET_NAME(ch));
-				AddKarma(ch, buf, GET_NAME(vict));
-				vict->set_remort(value);
-				SendMsgToGods(buf);
-			}
-			else {
-				SendMsgToChar(ch, "Не правильно указан реморт.\r\n");
-			}
-			break;
-		case 60: // флаг тестера
-			if (!str_cmp(val_arg, "off") || !str_cmp(val_arg, "выкл")) {
-				CLR_GOD_FLAG(vict, EGf::kAllowTesterMode);
-				PRF_FLAGS(vict).unset(EPrf::kTester); // обнулим реж тестер
-				sprintf(buf, "%s убрал флаг тестера для игрока %s", GET_NAME(ch), GET_NAME(vict));
-				mudlog(buf, BRF, kLvlImmortal, SYSLOG, true);
-			} else {
-				SET_GOD_FLAG(vict, EGf::kAllowTesterMode);
-				sprintf(buf, "%s установил флаг тестера для игрока %s", GET_NAME(ch), GET_NAME(vict));
-				mudlog(buf, BRF, kLvlImmortal, SYSLOG, true);
-				//			send_to_gods(buf);
-			}
-			break;
-		case 61: // флаг автобота
-		{
-			SET_OR_REMOVE(on, off, PLR_FLAGS(vict), EPlrFlag::kAutobot);
-			break;
-		}
-		case 62: vict->set_hryvn(value);
-			break;
-		case 63: // флаг скриптера
-			sprintf(buf, "%s", GET_NAME(ch));
-			if (!str_cmp(val_arg, "off") || !str_cmp(val_arg, "выкл")) {
-				PLR_FLAGS(vict).unset(EPlrFlag::kScriptWriter);
-				AddKarma(vict, "Снятие флага скриптера", buf);
-				sprintf(buf, "%s убрал флаг скриптера для игрока %s", GET_NAME(ch), GET_NAME(vict));
-				mudlog(buf, BRF, kLvlImmortal, SYSLOG, true);
-				return (1);
-			} else if (!str_cmp(val_arg, "on") || !str_cmp(val_arg, "вкл")) {
-				PLR_FLAGS(vict).set(EPlrFlag::kScriptWriter);
-				AddKarma(vict, "Установка флага скриптера", buf);
-				sprintf(buf, "%s установил  флаг скриптера для игрока %s", GET_NAME(ch), GET_NAME(vict));
-				mudlog(buf, BRF, kLvlImmortal, SYSLOG, true);
-				return (1);
-			} else {
-				SendMsgToChar(ch, "Значение может быть только on/off или вкл/выкл.\r\n");
-				return (0); // не пишем в пфайл ничего
-			}
-			break;
-		case 64: // флаг спамера
-		{
-			SET_OR_REMOVE(on, off, PLR_FLAGS(vict), EPlrFlag::kSpamer);
-			break;
-		}
-		case 65: { // спрятать отображение славы
-			if (!str_cmp(val_arg, "on") || !str_cmp(val_arg, "вкл")) {
-				GloryConst::glory_hide(vict, true);
-			} else {
-				GloryConst::glory_hide(vict, false);
-			}
-			break;
-		}
-		case 66: { // идентификатор чата телеграма
-
-			unsigned long int id = strtoul(val_arg, nullptr, 10);
-			if (!ch->IsNpc() && id != 0) {
-				sprintf(buf, "Telegram chat_id изменен с %lu на %lu\r\n", vict->player_specials->saved.telegram_id, id);
-				SendMsgToChar(buf, ch);
-				vict->setTelegramId(id);
-			} else
-				SendMsgToChar("Ошибка, указано неверное число или персонаж.\r\n", ch);
-			break;
-		}
-		case 67: vict->set_nogata(value);
-			break;
-		case 68: {
-			EPosition tmpval = (EPosition) value;
-			if (tmpval > EPosition::kDead && tmpval < EPosition::kLast) {
-				sprinttype(value, position_types, smallBuf);
-				sprintf(buf, "Для персонажа %s установлена позиция: %s.\r\n", GET_NAME(vict), smallBuf);
-				SendMsgToChar(buf, ch);
-				GET_POS(vict) = tmpval;
-			} else {
-				SendMsgToChar(ch, "Позиция может принимать значения от 1 до 8.\r\n");
-				return (0);
-			}
-			break;
-		}
-		case 69: // флаг скилл тестера
-			if (!str_cmp(val_arg, "off") || !str_cmp(val_arg, "выкл")) {
-				CLR_GOD_FLAG(vict, EGf::kSkillTester);
-				sprintf(buf, "%s убрал флаг &Rскилл тестера&n для игрока %s", GET_NAME(ch), GET_NAME(vict));
-				mudlog(buf, BRF, kLvlImmortal, SYSLOG, true);
-			} else {
-				SET_GOD_FLAG(vict, EGf::kSkillTester);
-				sprintf(buf, "%s установил флаг &Rскилл тестера&n для игрока %s", GET_NAME(ch), GET_NAME(vict));
-				mudlog(buf, BRF, kLvlImmortal, SYSLOG, true);
-			}
-			break;
-		case 70: //quest
-			vict->complete_quest(value);
-			SendMsgToChar(ch, "Добавил игроку %s выполнение квеста %d\r\n", GET_NAME(vict), value);
-			break;
-		default: SendMsgToChar("Не могу установить это!\r\n", ch);
-			return (0);
-	}
-
-	strcat(output, "\r\n");
-	SendMsgToChar(CAP(output), ch);
-	return (return_code);
-}
-
-void do_set(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/) {
-	CharData *vict = nullptr;
-	char field[kMaxInputLength], name[kMaxInputLength], val_arg[kMaxInputLength], OName[kMaxInputLength];
-	int mode, player_i = 0, retval;
-	char is_file = 0, is_player = 0;
-
-	half_chop(argument, name, buf);
-
-	if (!*name) {
-		strcpy(buf, "Возможные поля для изменения:\r\n");
-		for (int i = 0; set_fields[i].level; i++)
-			if (privilege::HasPrivilege(ch, std::string(set_fields[i].cmd), 0, 1))
-				sprintf(buf + strlen(buf), "%-15s%s", set_fields[i].cmd, (!((i + 1) % 5) ? "\r\n" : ""));
-		strcat(buf, "\r\n");
-		SendMsgToChar(buf, ch);
-		return;
-	}
-
-	if (!strcmp(name, "file")) {
-		is_file = 1;
-		half_chop(buf, name, buf);
-	} else if (!str_cmp(name, "player")) {
-		is_player = 1;
-		half_chop(buf, name, buf);
-	} else if (!str_cmp(name, "mob")) {
-		half_chop(buf, name, buf);
-	} else
-		is_player = 1;
-
-	half_chop(buf, field, buf);
-	strcpy(val_arg, buf);
-
-	if (!*name || !*field) {
-		SendMsgToChar("Usage: set [mob|player|file] <victim> <field> <value>\r\n", ch);
-		return;
-	}
-
-	CharData::shared_ptr cbuf;
-	// find the target
-	if (!is_file) {
-		if (is_player) {
-
-			if (!(vict = get_player_pun(ch, name, EFind::kCharInWorld))) {
-				SendMsgToChar("Нет такого игрока.\r\n", ch);
-				return;
-			}
-
-			// Запрет на злоупотребление командой SET на бессмертных
-			if (!GET_GOD_FLAG(ch, EGf::kDemigod)) {
-				if ((GetRealLevel(ch) <= GetRealLevel(vict)) && !(is_head(ch->get_name_str()))) {
-					SendMsgToChar("Вы не можете сделать этого.\r\n", ch);
-					return;
-				}
-			} else {
-				if (GetRealLevel(vict) >= kLvlImmortal) {
-					SendMsgToChar("Вы не можете сделать этого.\r\n", ch);
-					return;
-				}
-			}
-		} else    // is_mob
-		{
-			if (!(vict = get_char_vis(ch, name, EFind::kCharInWorld))
-				|| !vict->IsNpc()) {
-				SendMsgToChar("Нет такой твари Божьей.\r\n", ch);
-				return;
-			}
-		}
-	} else if (is_file)    // try to load the player off disk
-	{
-		if (get_player_pun(ch, name, EFind::kCharInWorld)) {
-			SendMsgToChar("Да разуй же глаза! Оно в сети!\r\n", ch);
-			return;
-		}
-
-		cbuf = std::make_unique<Player>();
-		if ((player_i = LoadPlayerCharacter(name, cbuf.get(), ELoadCharFlags::kFindId)) > -1) {
-			// Запрет на злоупотребление командой SET на бессмертных
-			if (!GET_GOD_FLAG(ch, EGf::kDemigod)) {
-				if (GetRealLevel(ch) <= GetRealLevel(cbuf) && !(is_head(ch->get_name_str()))) {
-					SendMsgToChar("Вы не можете сделать этого.\r\n", ch);
-					return;
-				}
-			} else {
-				if (GetRealLevel(cbuf) >= kLvlImmortal) {
-					SendMsgToChar("Вы не можете сделать этого.\r\n", ch);
-					return;
-				}
-			}
-			vict = cbuf.get();
-		} else {
-			SendMsgToChar("Нет такого игрока.\r\n", ch);
-			return;
-		}
-	}
-
-	// find the command in the list
-	size_t len = strlen(field);
-	for (mode = 0; *(set_fields[mode].cmd) != '\n'; mode++) {
-		if (!strncmp(field, set_fields[mode].cmd, len)) {
-			break;
-		}
-	}
-
-	// perform the set
-	strcpy(OName, GET_NAME(vict));
-	retval = perform_set(ch, vict, mode, val_arg);
-
-	// save the character if a change was made
-	if (retval && !vict->IsNpc()) {
-		if (retval == 2) {
-			RenamePlayer(vict, OName);
-		} else {
-			if (!is_file && !vict->IsNpc()) {
-				vict->save_char();
-			}
-			if (is_file) {
-				vict->save_char();
-				SendMsgToChar("Файл сохранен.\r\n", ch);
-			}
-		}
-	}
-
-	log("(GC) %s try to set: %s", GET_NAME(ch), argument);
-	imm_log("%s try to set: %s", GET_NAME(ch), argument);
 }
 
 namespace Mlist {
