+++ src/cmd_god/do_set.h
@@ -1,0 +1,13 @@
+//
+// Created by Sventovit on 05.09.2024.
+//
+
+#ifndef BYLINS_SRC_CMD_GOD_DO_SET_H_
+#define BYLINS_SRC_CMD_GOD_DO_SET_H_
+
+class CharData;
+void do_set(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/);
+
+#endif //BYLINS_SRC_CMD_GOD_DO_SET_H_
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
